Swapping Item and Items
In reviewing the naught gem, the plural is the container for the
different types. The singular constant is the container for the builder
methods. This brings Locabulary::Item(s) inline with the
Locabulary::Command(s)
