+++ lib/locabulary.rb
@@ -1,14 +1,13 @@
 require 'date'
 require 'json'
 require 'locabulary/exceptions'
-require 'locabulary/items'
+require 'locabulary/item'
 require 'locabulary/commands/build_ordered_hierarchical_tree_command'
 require 'locabulary/commands/active_items_for_command'
+require 'locabulary/commands/active_hierarchical_roots_command'
 
 # @since 0.1.0
 module Locabulary
-  DATA_DIRECTORY = File.expand_path("../../data/", __FILE__).freeze
-
   # @api private
   # @since 0.5.0
   #
@@ -29,12 +28,12 @@ module Locabulary
   # @api public
   # @since 0.1.0
   #
-  # Responsible for extracting a non-hierarchical sorted array of Locabulary::Item::Base objects for the given predicate_name.
+  # Responsible for extracting a non-hierarchical sorted array of Locabulary::Items::Base objects for the given predicate_name.
   #
   # @param [Hash] options
   # @option options [String] :predicate_name
   # @option options [Date] :as_of (Date.today)
-  # @return Array[<Locabulary::Item::Base>]
+  # @return Array[<Locabulary::Items::Base>]
   #
   # @see Locabulary::Commands::ActiveItemsForCommand
   def self.active_items_for(options = {})
@@ -50,7 +49,7 @@ module Locabulary
   # @param [Hash] options
   # @option options [String] :predicate_name
   # @option options [Date] :as_of (Date.today)
-  # @return [Array<Locabulary::Item::Base>] - the root nodes
+  # @return [Array<Locabulary::Items::Base>] - the root nodes
   def self.active_hierarchical_roots(options = {})
     Commands::ActiveHierarchicalRootsCommand.call(options)
   end
@@ -70,7 +69,7 @@ module Locabulary
     item = nil
     Utility.with_extraction_for(predicate_name) do |data|
       next unless data.fetch('term_label') == term_label
-      item = Items.build(data.merge('predicate_name' => predicate_name))
+      item = Item.build(data.merge('predicate_name' => predicate_name))
       break if Utility.data_is_active?(data, as_of)
     end
     return item unless item.nil?
@@ -103,7 +102,7 @@ module Locabulary
   # @param [Hash] options
   # @option options [String] :predicate_name
   # @option options [String] :as_of (Date.today)
-  # @return [Array<String>] an array of Locabuarly::Item::Base#term_label
+  # @return [Array<String>] an array of Locabuarly::Items::Base#term_label
   #
   # @see Locabulary.active_items_for
   def self.active_labels_for(options = {})
