+++ lib/locabulary/commands/active_items_for_command.rb
@@ -39,7 +39,7 @@ module Locabulary
       def initialize(options = {})
         @predicate_name = options.fetch(:predicate_name)
         @as_of = options.fetch(:as_of) { Date.today }
-        @builder = Items.builder_for(predicate_name: predicate_name)
+        @builder = Item.builder_for(predicate_name: predicate_name)
       end
 
       def call
