+++ lib/locabulary/facet_wrapper_for_item.rb
@@ -2,7 +2,7 @@ require 'forwardable'
 require 'delegate'
 
 module Locabulary
-  # A wrapper for a Locabulary::Item::Base that includes information from
+  # A wrapper for a Locabulary::Items::Base that includes information from
   # the SOLR Utility.
   class FacetWrapperForItem < SimpleDelegator
     # @api public
@@ -22,7 +22,7 @@ module Locabulary
       predicate_name = options.fetch(:predicate_name)
       faceted_node = options.fetch(:faceted_node)
       term_label = options.fetch(:term_label)
-      locabulary_item = Locabulary::Items.build(predicate_name: predicate_name, term_label: term_label, default_presentation_sequence: nil)
+      locabulary_item = Locabulary::Item.build(predicate_name: predicate_name, term_label: term_label, default_presentation_sequence: nil)
       new(faceted_node: faceted_node, locabulary_item: locabulary_item)
     end
