+++ spec/lib/locabulary/items/base_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
-require 'locabulary/item/base'
+require 'locabulary/items/base'
 
-RSpec.describe Locabulary::Item::Base do
+RSpec.describe Locabulary::Items::Base do
   context 'Comparable' do
     it 'sorts with nil and integers for presentation sequence' do
       item_1 = described_class.new(default_presentation_sequence: nil, term_label: 'Hello')
